+++ include/camera.hpp
@@ -43,7 +43,7 @@ class PerspectiveCamera : public Camera
 
 public:
     PerspectiveCamera(const Vector3f &center, const Vector3f &direction,
-                      const Vector3f &up, int imgW, int imgH, float angle) : Camera(center, direction, up, imgW, imgH)
+                      const Vector3f &up, int imgW, int imgH, double angle) : Camera(center, direction, up, imgW, imgH)
     {
         // angle is in radian.
         f_x = width / (2 * tan(angle / 2));
@@ -66,8 +66,8 @@ public:
     }
 
 private:
-    float f_x, f_y;
-    float angle;
+    double f_x, f_y;
+    double angle;
 };
 
 #endif //CAMERA_H
