+++ include/light.hpp
@@ -12,7 +12,7 @@ public:
 
     virtual ~Light() = default;
 
-    virtual void getIllumination(const Vector3f &p, Vector3f &dir, Vector3f &col, float &distanceToLight) const = 0;
+    virtual void getIllumination(const Vector3f &p, Vector3f &dir, Vector3f &col, double &distanceToLight) const = 0;
 };
 
 class DirectionalLight : public Light
@@ -29,7 +29,7 @@ public:
     ~DirectionalLight() override = default;
     ///@param p unsed in this function
     ///@param distanceToLight not well defined because it's not a point light
-    virtual void getIllumination(const Vector3f &p, Vector3f &dir, Vector3f &col, float &distanceToLight) const
+    virtual void getIllumination(const Vector3f &p, Vector3f &dir, Vector3f &col, double &distanceToLight) const
     {
         // the direction to the light is the opposite of the
         // direction of the directional light source
@@ -46,7 +46,7 @@ private:
 class PointLight : public Light
 {
 public:
-    PointLight(const Vector3f &p, const Vector3f &c, float fall)
+    PointLight(const Vector3f &p, const Vector3f &c, double fall)
     {
         position = p;
         color = c;
@@ -55,7 +55,7 @@ public:
 
     ~PointLight() override = default;
 
-    virtual void getIllumination(const Vector3f &p, Vector3f &dir, Vector3f &col, float &distanceToLight) const
+    virtual void getIllumination(const Vector3f &p, Vector3f &dir, Vector3f &col, double &distanceToLight) const
     {
         // the direction to the light is the opposite of the
         // direction of the directional light source
@@ -66,7 +66,7 @@ public:
     }
 
 private:
-    float falloff;
+    double falloff;
     Vector3f position;
     Vector3f color;
 };
