+++ include/sphere.hpp
@@ -14,7 +14,7 @@ public:
     }
 
     Sphere(const Vector3f &center,
-           float radius,
+           double radius,
            Material *material) : Object3D(material),
                                  center(center),
                                  radius(radius),
@@ -25,7 +25,7 @@ public:
 
     ~Sphere() override = default;
 
-    bool intersect(const Ray &r, Hit &h, float t_min) override
+    bool intersect(const Ray &r, Hit &h, double t_min) override
     {
         const auto &origin = r.getOrigin();
         const auto &direction = r.getNormalizedDirection();
@@ -62,7 +62,7 @@ protected:
         OUTSIDE,
     };
 
-    Relation getRelation(float square_length)
+    Relation getRelation(double square_length)
     {
         if (square_length > radius)
         {
@@ -79,8 +79,8 @@ protected:
     }
 
     Vector3f center;
-    float radius;
-    float radius_square;
+    double radius;
+    double radius_square;
 };
 
 #endif
