+++ src/noise.cpp
@@ -9,7 +9,7 @@ Vector3f Noise::getColor(const Vector3f &pos)
 }
 
 Noise::Noise(int _octaves, const Vector3f &color1,
-			 const Vector3f &color2, float freq, float amp) : octaves(_octaves), frequency(freq), amplitude(amp)
+			 const Vector3f &color2, double freq, double amp) : octaves(_octaves), frequency(freq), amplitude(amp)
 {
 	color[0] = color1;
 	color[1] = color2;
